requirements.txt
flask
langchain
langchain-core
langchain-community
langchain-chroma
langchain-huggingface
langchain-text-splitters
faiss-cpu==1.9.0
sentence-transformers
python-dotenv
requests
google-generativeai
numpy
protobuf
